refactor: inputView 검증 로직 -> Util로 이동
